+++ .idea/dictionaries/project.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="project">
+    <words>
+      <w>datagroup</w>
+    </words>
+  </dictionary>
+</component>
